+++ test/test_validation.py
@@ -6,9 +6,12 @@ import os
 import numpy as np
 import unittest
 
-from atoml.cross_validation import Hierarchy
+from atoml.cross_validation import Hierarchy, k_fold
+from atoml.cross_validation.k_fold_cv import write_split, read_split
 from atoml.regression import RidgeRegression
 from common import get_data
+
+wkdir = os.getcwd()
 
 
 def predict(train_features, train_targets, test_features, test_targets):
@@ -36,7 +39,7 @@ def predict(train_features, train_targets, test_features, test_targets):
     return data
 
 
-class TestHierarchy(unittest.TestCase):
+class TestValidation(unittest.TestCase):
     """Test out the hierarchy cv."""
 
     def test_hierarchy(self):
@@ -44,22 +47,41 @@ class TestHierarchy(unittest.TestCase):
         # Define the hierarchy cv class method.
         train_features, train_targets, test_features, test_targets = get_data()
 
-        hv = Hierarchy(db_name='test.sqlite', table='FingerVector',
-                       file_name='hierarchy')
+        hv = Hierarchy(
+            db_name='{}/test.sqlite'.format(wkdir), file_name='hierarchy')
         hv.todb(features=train_features, targets=train_targets)
         # Split the data into subsets.
         split = hv.split_index(min_split=5, max_split=25)
-        self.assertTrue(len(split) == 6)
+        self.assertEqual(len(split), 6)
         # Load data back in from save file.
         ind = hv.load_split()
-        self.assertTrue(len(ind) == 6)
+        self.assertEqual(len(ind), len(split))
 
         # Make the predictions for each subset.
         pred = hv.split_predict(index_split=ind, predict=predict)
         self.assertTrue(len(pred[0]) == 14 and len(pred[1]) == 14)
 
-        os.remove('hierarchy.pickle')
-        os.remove('test.sqlite')
+    def test_kfold(self):
+        """Test some cross-validation."""
+        features, targets, _, _ = get_data()
+        f, t = k_fold(features, targets, nsplit=5)
+        self.assertTrue(len(f) == 5 and len(t) == 5)
+        for s in f:
+            self.assertEqual(np.shape(s), (9, 100))
+        f, t = k_fold(features, targets, nsplit=4, fix_size=5)
+        self.assertTrue(len(f) == 4 and len(t) == 4)
+        for s in f:
+            self.assertEqual(np.shape(s), (5, 100))
+
+        write_split(features=f, targets=t, fname='cvsave', fformat='pickle')
+        f1, t1 = read_split(fname='cvsave', fformat='pickle')
+        self.assertEqual(len(f1), len(f))
+        self.assertEqual(len(t1), len(t))
+
+        write_split(features=f, targets=t, fname='cvsave', fformat='json')
+        f1, t1 = read_split(fname='cvsave', fformat='pickle')
+        self.assertEqual(len(f1), len(f))
+        self.assertEqual(len(t1), len(t))
 
 
 if __name__ == '__main__':