+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Version 0.3.1 (February 2018)
+
+-   Added a parallel version of the greedy feature selection. **Python3 only!**
+-   Updated the k-fold cross-validation function to handle features and targets explicitly.
+-   Added some basic read/write functionality to the k-fold CV.
+-   A number of minor bugs have been fixed.
+
 # Version 0.3.0 (February 2018)
 
 -   Update the fingerprint generator functions so there is now a `FeatureGenerator` class that wraps round all type specific generators.